// src/assets/styles/interface.ts
export interface IDefaultTheme {
  appColors?: {
    backgroundColor?: string;
    backgroundColorDarken?: string;
    backgroundColorLighter?: string;
    primaryColorDarken?: string;
    primaryColorLighter?: string;
    accentColor?: string;
    textColorLightGray?: string;
    textColorDarkGray?: string;
    errorColor?: string;
    whiteColor?: string;
    blackColor?: string;
  };
}